requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
Django==5.2.8
djangorestframework==3.16.1
dnspython==2.8.0
drf-yasg==1.21.11
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.121.2
gevent==25.9.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
kombu==5.6.1
lxml==6.0.2
packaging==25.0
passlib==1.7.4
pillow==12.0.0
phonenumbers
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
reportlab==4.4.6
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.3
twilio
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
zope.event==6.1
zope.interface==8.1.1
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0
pymysql==1.1.1
django-cors-headers==4.3.1
